LHC-1820

+++ modules/dxp/apps/osb/osb-customer/osb-customer-admin/osb-customer-admin-web/src/main/java/com/liferay/osb/customer/admin/web/internal/portlet/AdminPortlet.java
@@ -221,6 +221,16 @@ public class AdminPortlet extends MVCPortlet {
 		long productEntryId = ParamUtil.getLong(
 			actionRequest, "productEntryId");
 
+		ProductEntry productEntry = _productEntryLocalService.getProductEntry(
+			productEntryId);
+
+		Product product = _productWebService.fetchProduct(
+			productEntry.getKoroneikiProductKey());
+
+		if (product != null) {
+			throw new RequiredProductEntryException();
+		}
+
 		_productEntryLocalService.deleteProductEntry(productEntryId);
 	}
